refactor: rename draw elements
