+++ src/draw.rs
@@ -15,13 +15,6 @@ pub enum EventResult {
     Ignored,
 }
 
-// not call draw themselves. Then we can pass custom variables through the draw call, should remove
-// FIX: rewrite such that StackLayout and AdjacentLayout only generates layouts and
-// all uses of RefCell code. This would also allow more flexibility. Perhaps DrawableComponent
-// should be an enum, however less flexibility, but would allow for nice callback, since pop_layer
-// would know which componenent is which.
-// ie: viewer.draw(&mut app, layout, task_list.selected, completed_list.selected)
-
 /// A component that is able to be drawn on the screen.
 pub trait DrawableComponent {
     /// Draws the component onto the [[Drawer]]
@@ -45,26 +38,22 @@ pub trait DrawableComponent {
 // How does this even work, mind blown, wait does it give back ownership when it's done, if so
 // that's just really fucking cool.
 pub struct Drawer<'a, 'b, 'c> {
-    backend: &'a mut DrawBackend<'b, 'c>,
+    backend: &'a mut DrawFrame<'b, 'c>,
 }
 
 impl Drawer<'_, '_, '_> {
-    pub fn new<'a, 'b, 'c>(backend: &'a mut DrawBackend<'b, 'c>) -> Drawer<'a, 'b, 'c> {
+    pub fn new<'a, 'b, 'c>(backend: &'a mut DrawFrame<'b, 'c>) -> Drawer<'a, 'b, 'c> {
         Drawer { backend }
     }
 
    pub fn draw_component(&mut self, app: &App, drawable: &dyn DrawableComponent, draw_area: Rect) {
-        let mut render = Drawer {
-            backend: self.backend,
-        };
-        drawable.draw(app, draw_area, &mut render);
+        drawable.draw(app, draw_area, self);
     }
 
     pub fn draw_widget<'a, T: Widget>(&mut self, widget: T, draw_area: Rect) {
         self.backend.draw_widget(widget, draw_area);
     }
 
-    // TODO: does not acknowledge the area
     pub fn draw_stateful_widget<T: StatefulWidget>(
         &mut self,
         widget: T,
@@ -79,14 +68,14 @@ impl Drawer<'_, '_, '_> {
     }
 }
 
-pub enum DrawBackend<'a, 'b> {
+pub enum DrawFrame<'a, 'b> {
     CrosstermRenderer(&'a mut Frame<'b, CrosstermBackend<Stdout>>),
 }
 
-impl DrawBackend<'_, '_> {
+impl DrawFrame<'_, '_> {
     fn draw_widget<T: Widget>(&mut self, widget: T, draw_area: Rect) {
         match self {
-            DrawBackend::CrosstermRenderer(f) => {
+            DrawFrame::CrosstermRenderer(f) => {
                 f.render_widget(widget, draw_area);
             }
         }
@@ -99,7 +88,7 @@ impl DrawBackend<'_, '_> {
         draw_area: Rect,
     ) {
         match self {
-            DrawBackend::CrosstermRenderer(f) => {
+            DrawFrame::CrosstermRenderer(f) => {
                 f.render_stateful_widget(widget, draw_area, state);
             }
         }
@@ -107,7 +96,7 @@ impl DrawBackend<'_, '_> {
 
     fn set_cursor(&mut self, x: u16, y: u16) {
         match self {
-            DrawBackend::CrosstermRenderer(f) => {
+            DrawFrame::CrosstermRenderer(f) => {
                 f.set_cursor(x, y);
             }
         }
